Crud expense v16

+++ loan-tracking-backend/src/main/java/main/loantrackingbackend/service/impl/GroupMemberImpl.java
@@ -7,15 +7,15 @@ import main.loantrackingbackend.entity.Group;
 import main.loantrackingbackend.entity.GroupMember;
 import main.loantrackingbackend.entity.Person;
 import main.loantrackingbackend.entity.embedabble.GroupMemberId;
+import main.loantrackingbackend.enums.PaymentStatus;
 import main.loantrackingbackend.exception.ResourceNotFoundException;
 import main.loantrackingbackend.mapper.GroupMemberMapper;
 import main.loantrackingbackend.mapper.PersonMapper;
-import main.loantrackingbackend.repository.GroupMemberRepository;
-import main.loantrackingbackend.repository.GroupRepository;
-import main.loantrackingbackend.repository.PersonRepository;
+import main.loantrackingbackend.repository.*;
 import main.loantrackingbackend.service.GroupMemberService;
 import org.springframework.stereotype.Service;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -23,12 +23,15 @@ import java.util.stream.Collectors;
 @AllArgsConstructor
 public class GroupMemberImpl implements GroupMemberService {
 
+    private final GroupExpenseRepository groupExpenseRepository;
+    private final PaymentAllocationRepository paymentAllocationRepository;
     private PersonRepository personRepository;
     private GroupRepository groupRepository;
     private GroupMemberRepository groupMemberRepository;
 
     @Override
     public GroupMemberDto addMember(Long groupId, Long personId) {
+
         Group group = groupRepository.findById(groupId)
                 .orElseThrow(() -> new ResourceNotFoundException("Group not found"));
 
@@ -37,8 +40,18 @@ public class GroupMemberImpl implements GroupMemberService {
 
         GroupMemberId id = new GroupMemberId(groupId, personId);
 
+        List<PaymentStatus> activeStatuses = Arrays.asList(
+                PaymentStatus.UNPAID,
+                PaymentStatus.PARTIALLY_PAID
+        );
+
         if (groupMemberRepository.existsById(id)) {
             throw new IllegalStateException("Person already in group");
+        }
+
+        if (groupExpenseRepository.existsByGroupBorrowerAndStatusIn(group, activeStatuses)) {
+            throw new IllegalArgumentException("Cannot add member to group " + group.getGroupName() + " since group is " +
+                    " associated with ongoing expense.");
         }
 
         GroupMember member = new GroupMember();
@@ -48,17 +61,28 @@ public class GroupMemberImpl implements GroupMemberService {
 
         GroupMember savedGroupMember = groupMemberRepository.save(member);
 
+        person.getGroupMembers().add(savedGroupMember);
+        personRepository.save(person);
+
         return GroupMemberMapper.mapToGroupMemberDto(savedGroupMember);
     }
 
     @Override
-    public void removeMember(Long groupId, Long personId) {
+    public String removeMember(Long groupId, Long personId) {
         GroupMemberId id = new GroupMemberId(groupId, personId);
 
         GroupMember member = groupMemberRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Member not found"));
 
+        if (paymentAllocationRepository.existsByGroupMemberBorrower(member.getPerson()))
+            return "Cannot delete group member " + member.getPerson().getFirstName() + member.getPerson().getLastName() +
+                    " with id " + member.getPerson().getPersonId() +
+                    " since they are involved with group expense/s. Delete those expense first.";
+
         groupMemberRepository.delete(member);
+        return "Successfully deleted group member " + member.getPerson().getFirstName()
+                + member.getPerson().getLastName() +
+                " with id " + member.getPerson().getPersonId();
     }
 
     @Override
