+++ loan-tracking-backend/src/main/java/main/loantrackingbackend/service/impl/GroupServiceImpl.java
@@ -3,12 +3,17 @@ package main.loantrackingbackend.service.impl;
 import lombok.AllArgsConstructor;
 import main.loantrackingbackend.dto.GroupDto;
 import main.loantrackingbackend.entity.Group;
+import main.loantrackingbackend.enums.PaymentStatus;
+import main.loantrackingbackend.exception.DuplicateResourceException;
 import main.loantrackingbackend.exception.ResourceNotFoundException;
 import main.loantrackingbackend.mapper.GroupMapper;
+import main.loantrackingbackend.repository.GroupExpenseRepository;
+import main.loantrackingbackend.repository.GroupMemberRepository;
 import main.loantrackingbackend.repository.GroupRepository;
 import main.loantrackingbackend.service.GroupService;
 import org.springframework.stereotype.Service;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -16,11 +21,18 @@ import java.util.stream.Collectors;
 @AllArgsConstructor
 public class GroupServiceImpl implements GroupService {
 
+    private final GroupExpenseRepository groupExpenseRepository;
     private GroupRepository groupRepository;
+    private GroupMemberRepository groupMemberRepository;
 
     @Override
     public GroupDto createGroup(GroupDto groupDto) {
+        if (groupRepository.existsByGroupName(groupDto.getGroupName())) {
+            throw new DuplicateResourceException("Group with name " + groupDto.getGroupName() + " already exists");
+        }
+
         Group group = GroupMapper.mapToGroup(groupDto);
+
         Group saveGroup = groupRepository.save(group);
 
         return GroupMapper.mapToGroupDto(saveGroup);
@@ -43,6 +55,10 @@ public class GroupServiceImpl implements GroupService {
 
     @Override
     public GroupDto updateGroup(Long groupId, GroupDto groupDto) {
+        if (groupRepository.existsByGroupName(groupDto.getGroupName())) {
+            throw new DuplicateResourceException("Group with name " + groupDto.getGroupName() + " already exists");
+        }
+
         Group group = groupRepository.findById(groupId)
                 .orElseThrow(() -> new ResourceNotFoundException("Group does not exist with id: " + groupId));
 
@@ -54,10 +70,25 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public void deleteGroup(Long groupId) {
+    public String deleteGroup(Long groupId) {
         Group group = groupRepository.findById(groupId)
                 .orElseThrow(() -> new ResourceNotFoundException("Group does not exist with id: " + groupId));
 
+        if (groupExpenseRepository.existsByGroupBorrower(group)) {
+            return "Cannot delete Group " + group.getGroupName() + " with group id: " + groupId +
+                    "because it is associated with expense/s. Delete those expense first.";
+        }
+
         groupRepository.delete(group);
+        return "Group " + group.getGroupName() + "with group id: " + groupId + " has been deleted";
+    }
+
+    @Override
+    public long countMembers(Long groupId) {
+        if (!groupRepository.existsById(groupId)) {
+            throw new ResourceNotFoundException("Group not found with id: " + groupId);
+        }
+
+        return groupMemberRepository.countByGroup_GroupId(groupId);
     }
 }
